refactor: classify functions to util.py and img.py
