+++ mod/img.py
@@ -19,9 +19,27 @@ def load_images(images_path):
             glob.glob(os.path.join(images_path, "*.png")) + \
             glob.glob(os.path.join(images_path, "*.jpeg"))
 
-def output_time():
-    loc_dt = datetime.datetime.today() 
-    time_del = datetime.timedelta(hours=0)
-    new_dt = loc_dt + time_del 
-    datetime_format = new_dt.strftime("%Y-%m-%d_%H-%M-%S")
-    return datetime_format
+def get_imginfo(img):
+    img_shape = img.shape[:2] # (height, width)
+    size = (img_shape[1], img_shape[0]) # (width, height)
+    return size
+
+def convert_bboxes(labeling_data, img_size):
+    
+    _bboxes = list()
+    left, top, width, height = labeling_data[1:]
+
+    x =  img_size[0] * left       
+    y =  img_size[1] * top        
+    w =  (img_size[0] * width) / 2      
+    h =  (img_size[1] * height) / 2
+
+    xmin = x - w
+    ymin = y - h
+    xmax = x + w
+    ymax = y + h 
+
+    list_bbox = [xmin, ymin, xmax, ymax, labeling_data[0]]
+    
+    _bboxes.append(list_bbox)
+    return _bboxes
